requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-core
chromadb
openai
pypdf
docx2txt
unstructured
python-pptx
pandas
pillow
pytesseract
python-multipart